Rename components folder (#24)
* Renamed components folder

* Update App.tsx

* Temporary name for components folder

* Renamed folder
