+++ frontend/src/components/SimpleGetRequestTemplate.tsx
